Move tests to their dedicated file
