+++ core/unit_test/CMakeLists.txt
@@ -198,6 +198,7 @@ foreach(Tag Threads;Serial;OpenMP;Cuda;HPX;OpenMPTarget;OpenACC;HIP;SYCL)
         MDRange_f
         MDRange_g
         MDRangePolicyConstructors
+        MDRangePolicyLimits
         MDRangeReduce
         MDSpan
         MDSpanAtomicAccessor
